+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [unreleased]
+
+### Added
+
+- Added assertResponseRegExp to IntegrationTestTrait
+- Added assertResponseNotRegExp to IntegrationTestTrait
+
 ## [3.4.1] - 2020-08-24
 
 ### Fixed